+++ CHANGELOG.md
@@ -5,6 +5,24 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/williamiommi/sanity-plugin-icon-manager/compare/v2.2.1...v2.2.2) (2024-12-15)
+
+### Bug Fixes
+
+- better popover management with portal and on blur ([9bce10f](https://github.com/williamiommi/sanity-plugin-icon-manager/commit/9bce10f872ccd424fb27acfa7e941a692f2ff7b5))
+
+## [2.2.1](https://github.com/williamiommi/sanity-plugin-icon-manager/compare/v2.2.0...v2.2.1) (2024-12-15)
+
+### Bug Fixes
+
+- added a 20px max width to input icon ([20c2c6e](https://github.com/williamiommi/sanity-plugin-icon-manager/commit/20c2c6e75bcbb673b7572fdc6baca76d87ecb703))
+
+## [2.2.0](https://github.com/williamiommi/sanity-plugin-icon-manager/compare/v2.1.0...v2.2.0) (2024-12-14)
+
+### Features
+
+- **simple-ui:** basic UI ([4f2e81a](https://github.com/williamiommi/sanity-plugin-icon-manager/commit/4f2e81a8aa84d011bbbf936eb5d8242a231c0ecf))
+
 ## [2.1.0](https://github.com/williamiommi/sanity-plugin-icon-manager/compare/v2.0.2...v2.1.0) (2024-09-03)
 
 ### Features